+++ src/Symfony/Component/Lock/Store/RedisStore.php
@@ -74,17 +74,6 @@ class RedisStore implements PersistingStoreInterface
 
     /**
      * {@inheritdoc}
-     *
-     * @deprecated since Symfony 4.4.
-     */
-    public function waitAndSave(Key $key)
-    {
-        @trigger_error(sprintf('%s() is deprecated since Symfony 4.4 and will be removed in Symfony 5.0.', __METHOD__), E_USER_DEPRECATED);
-        throw new InvalidArgumentException(sprintf('The store "%s" does not supports blocking locks.', \get_class($this)));
-    }
-
-    /**
-     * {@inheritdoc}
      */
     public function putOffExpiration(Key $key, $ttl)
     {