+++ src/Symfony/Component/Lock/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 * `Factory` has been removed, use `LockFactory` instead.
 * `StoreInterface` has been removed, use `BlockingStoreInterface` and `PersistingStoreInterface` instead.
+* removed the `waitAndSave()` method from `CombinedStore`, `MemcachedStore`, `RedisStore`, and `ZookeeperStore`
 
 4.4.0
 -----